refactor(workspace): tidy up after the automatic grouping units
Three parallel reviewers over the branch diff; the findings worth taking:

Route every manual grouping path through note_manual_tab_placement
instead of writing tab.placed_by_automation inline. Its own doc calls it
the only automation state a manual placement writes, and one caller in
the same file already used it, but three others each restated the body.
The one remaining inline write iterates the tab list mutably and cannot
call the helper without restructuring.

Skip resolving a project key for a pinned tab. A pinned tab is never
grouped, so the repository-cache walk and the freshly built list of every
group's key produced a value the reconcile pass discarded -- on every cd,
for the life of the window. The anchor is still refreshed, so unpinning
still resolves against current state.

Keep the Appearance page's action enum private and expose one narrow
constructor for the toggle the tab menu dispatches. Making the whole
30-variant enum crate-visible to build a single value coupled the tab
menu to the settings page's internal representation.

Share one contiguity assertion between the three auto-grouping test
modules, which had byte-identical copies. Three copies let a weakened
assertion silently disable the check in a single file.
