+++ app/src/workspace/view/auto_grouping_wiring_tests.rs
@@ -144,23 +144,6 @@ fn tab_index_of(workspace: &Workspace, pane_group_id: EntityId) -> usize {
         .expect("tab still exists")
 }
 
-/// Group members occupying a contiguous run is a convention nothing enforces at
-/// runtime, so every case that moves a tab asserts it.
-fn assert_groups_contiguous(workspace: &Workspace) {
-    for group_id in workspace.tab_groups.keys() {
-        let indices: Vec<usize> = group_member_indices(&workspace.tabs, *group_id).collect();
-        let Some(&first) = indices.first() else {
-            continue;
-        };
-        let last = indices[indices.len() - 1];
-        assert_eq!(
-            last - first + 1,
-            indices.len(),
-            "group {group_id:?} members are not a contiguous run: {indices:?}"
-        );
-    }
-}
-
 #[test]
 fn directory_change_between_two_non_git_directories_reconciles() {
     let _grouped_tabs_guard = FeatureFlag::GroupedTabs.override_enabled(true);
